Moved event managers into Loop/Events namespace.
