+++ src/Loop/Events/EventFactoryInterface.php
@@ -1,14 +1,14 @@
 <?php
 namespace Icicle\Loop\Events;
 
-use Icicle\Loop\Manager\ImmediateManagerInterface;
-use Icicle\Loop\Manager\SocketManagerInterface;
-use Icicle\Loop\Manager\TimerManagerInterface;
+use Icicle\Loop\Events\Manager\ImmediateManagerInterface;
+use Icicle\Loop\Events\Manager\SocketManagerInterface;
+use Icicle\Loop\Events\Manager\TimerManagerInterface;
 
 interface EventFactoryInterface
 {
     /**
-     * @param   \Icicle\Loop\Manager\SocketManagerInterface $manager
+     * @param   \Icicle\Loop\Events\Manager\SocketManagerInterface $manager
      * @param   resource $resource Socket resource.
      * @param   callable $callback Callback function invoked when data is available on the socket.
      *
@@ -17,7 +17,7 @@ interface EventFactoryInterface
     public function socket(SocketManagerInterface $manager, $resource, callable $callback);
     
     /**
-     * @param   \Icicle\Loop\Manager\TimerManagerInterface $manager
+     * @param   \Icicle\Loop\Events\Manager\TimerManagerInterface $manager
      * @param   callable $callback Callback function invoked after the interval elapses.
      * @param   int|float $interval Timer interval.
      * @param   bool $periodic Set to true to repeat the timer every interval seconds, false for a one-time timer.
@@ -28,7 +28,7 @@ interface EventFactoryInterface
     public function timer(TimerManagerInterface $manager, callable $callback, $interval, $periodic = false, array $args = null);
     
     /**
-     * @param   \Icicle\Loop\Manager\ImmediateManagerInterface $manager
+     * @param   \Icicle\Loop\Events\Manager\ImmediateManagerInterface $manager
      * @param   callable $callback Callback function to be invoked.
      * @param   mixed[]|null $args Arguments to pass to the callback function.
      *
