+++ src/Loop/Events/Manager/Event/PollManager.php
@@ -1,17 +1,17 @@
 <?php
-namespace Icicle\Loop\Manager\Event;
+namespace Icicle\Loop\Events\Manager\Event;
 
 use Event;
 use EventBase;
 use Icicle\Loop\Events\SocketEventInterface;
 
-class AwaitManager extends SocketManager
+class PollManager extends SocketManager
 {
     /**
      * @inheritdoc
      */
     protected function createEvent(EventBase $base, SocketEventInterface $socket, callable $callback)
     {
-        return new Event($base, $socket->getResource(), Event::WRITE, $callback, $socket);
+        return new Event($base, $socket->getResource(), Event::READ, $callback, $socket);
     }
 }
