+++ src/Loop/Events/Manager/Libevent/SocketManager.php
@@ -1,20 +1,21 @@
 <?php
-namespace Icicle\Loop\Manager\Event;
+namespace Icicle\Loop\Events\Manager\Libevent;
 
 use Event;
 use EventBase;
 use Icicle\Loop\Events\EventFactoryInterface;
+use Icicle\Loop\Events\Manager\SocketManagerInterface;
 use Icicle\Loop\Events\SocketEventInterface;
 use Icicle\Loop\Exception\FreedException;
 use Icicle\Loop\Exception\ResourceBusyException;
-use Icicle\Loop\Manager\SocketManagerInterface;
 
 abstract class SocketManager implements SocketManagerInterface
 {
     const MIN_TIMEOUT = 0.001;
+    const MICROSEC_PER_SEC = 1e6;
 
     /**
-     * @var \EventBase
+     * @var resource
      */
     private $base;
     
@@ -24,7 +25,7 @@ abstract class SocketManager implements SocketManagerInterface
     private $factory;
     
     /**
-     * @var \Event[]
+     * @var resource[]
      */
     private $events = [];
     
@@ -34,26 +35,31 @@ abstract class SocketManager implements SocketManagerInterface
     private $sockets = [];
     
     /**
+     * @var bool[]
+     */
+    private $pending = [];
+    
+    /**
      * @var callable
      */
     private $callback;
     
     /**
-     * Creates an Event object on the given EventBase for the SocketEventInterface.
+     * Creates an event resource on the given event base for the SocketEventInterface.
      *
-     * @param   \EventBase $base
+     * @param   resource $base Event base resource.
      * @param   \Icicle\Loop\Events\SocketEventInterface $event
      * @param   callable $callback
      *
-     * @return  \Event
+     * @return  resource Event resource.
      */
-    abstract protected function createEvent(EventBase $base, SocketEventInterface $event, callable $callback);
+    abstract protected function createEvent($base, SocketEventInterface $event, callable $callback);
     
     /**
      * @param   \Icicle\Loop\Events\EventFactoryInterface $factory
-     * @param   \EventBase $base
+     * @param   resource $base
      */
-    public function __construct(EventFactoryInterface $factory, EventBase $base)
+    public function __construct(EventFactoryInterface $factory, $base)
     {
         $this->factory = $factory;
         $this->base = $base;
@@ -67,7 +73,7 @@ abstract class SocketManager implements SocketManagerInterface
     public function __destruct()
     {
         foreach ($this->events as $event) {
-            $event->free();
+            event_free($event);
         }
     }
     
@@ -76,8 +82,8 @@ abstract class SocketManager implements SocketManagerInterface
      */
     public function isEmpty()
     {
-        foreach ($this->events as $event) {
-            if ($event->pending) {
+        foreach ($this->pending as $pending) {
+            if ($pending) {
                 return false;
             }
         }
@@ -93,10 +99,13 @@ abstract class SocketManager implements SocketManagerInterface
         $id = (int) $resource;
         
         if (isset($this->sockets[$id])) {
-            throw new ResourceBusyException('A poll has already been created for that resource.');
+            throw new ResourceBusyException('An event has already been created for that resource.');
         }
         
-        return $this->sockets[$id] = $this->factory->socket($this, $resource, $callback);
+        $this->sockets[$id] = $this->factory->socket($this, $resource, $callback);
+        $this->pending[$id] = false;
+        
+        return $this->sockets[$id];
     }
     
     /**
@@ -114,17 +123,19 @@ abstract class SocketManager implements SocketManagerInterface
             $this->events[$id] = $this->createEvent($this->base, $socket, $this->callback);
         }
 
+        $this->pending[$id] = true;
+
         if (null === $timeout) {
-            $this->events[$id]->add();
+            event_add($this->events[$id]);
             return;
         }
-        
+
         $timeout = (float) $timeout;
         if (self::MIN_TIMEOUT > $timeout) {
             $timeout = self::MIN_TIMEOUT;
         }
 
-        $this->events[$id]->add($timeout);
+        event_add($this->events[$id], $timeout * self::MICROSEC_PER_SEC);
     }
     
     /**
@@ -135,7 +146,8 @@ abstract class SocketManager implements SocketManagerInterface
         $id = (int) $socket->getResource();
         
         if (isset($this->sockets[$id], $this->events[$id]) && $socket === $this->sockets[$id]) {
-            $this->events[$id]->del();
+            event_del($this->events[$id]);
+            $this->pending[$id] = false;
         }
     }
     
@@ -146,7 +158,7 @@ abstract class SocketManager implements SocketManagerInterface
     {
         $id = (int) $socket->getResource();
         
-        return isset($this->sockets[$id], $this->events[$id]) && $socket === $this->sockets[$id] && $this->events[$id]->pending;
+        return isset($this->sockets[$id], $this->pending[$id]) && $socket === $this->sockets[$id] && $this->pending[$id];
     }
     
     /**
@@ -158,9 +170,10 @@ abstract class SocketManager implements SocketManagerInterface
         
         if (isset($this->sockets[$id]) && $socket === $this->sockets[$id]) {
             unset($this->sockets[$id]);
+            unset($this->pending[$id]);
             
             if (isset($this->events[$id])) {
-                $this->events[$id]->free();
+                event_free($this->events[$id]);
                 unset($this->events[$id]);
             }
         }
@@ -182,11 +195,12 @@ abstract class SocketManager implements SocketManagerInterface
     public function clear()
     {
         foreach ($this->events as $event) {
-            $event->free();
+            event_free($event);
         }
         
         $this->events = [];
         $this->sockets = [];
+        $this->pending = [];
     }
     
     /**
@@ -195,7 +209,8 @@ abstract class SocketManager implements SocketManagerInterface
     protected function createCallback()
     {
         return function ($resource, $what, SocketEventInterface $socket) {
-            $socket->call(0 !== (Event::TIMEOUT & $what));
+            $this->pending[(int) $resource] = false;
+            $socket->call(0 !== (EV_TIMEOUT & $what));
         };
     }
 }
