+++ src/Loop/LibeventLoop.php
@@ -2,12 +2,12 @@
 namespace Icicle\Loop;
 
 use Icicle\Loop\Events\EventFactoryInterface;
+use Icicle\Loop\Events\Manager\Libevent\AwaitManager;
+use Icicle\Loop\Events\Manager\Libevent\PollManager;
+use Icicle\Loop\Events\Manager\Libevent\TimerManager;
+use Icicle\Loop\Events\Manager\SocketManagerInterface;
+use Icicle\Loop\Events\Manager\TimerManagerInterface;
 use Icicle\Loop\Exception\UnsupportedException;
-use Icicle\Loop\Manager\Libevent\AwaitManager;
-use Icicle\Loop\Manager\Libevent\PollManager;
-use Icicle\Loop\Manager\Libevent\TimerManager;
-use Icicle\Loop\Manager\SocketManagerInterface;
-use Icicle\Loop\Manager\TimerManagerInterface;
 
 /**
  * Uses the libevent extension to poll sockets for I/O and create timers.
